Add course list

+++ src/main/java/com/jhuoose/monadic/models/Lesson.java
@@ -3,11 +3,11 @@ package com.jhuoose.monadic.models;
 import java.util.ArrayList;
 
 public class Lesson {
-    double ID;
+    int ID;
     ArrayList<LessonElement> lessonElements;
     String description;
 
-    public Lesson(double ID, ArrayList<LessonElement> lessonElements, String description) {
+  public Lesson(double ID, ArrayList<LessonElement> lessonElements, String description) {
         this.ID = ID;
         this.lessonElements = lessonElements;
         this.description = description;
@@ -17,7 +17,7 @@ public class Lesson {
         return ID;
     }
 
-    public void setID(double ID) {
+    public void setID(int ID) {
         this.ID = ID;
     }
